Update: Wheelspeed Works
Now this sensor hub reads the wheel speed of a pair of front or rear wheels.

Verified to capture the PWM inputs of TIM3 and TIM15 and convert them to proper RPM Values. The measured values coincided with 

The pulse issue was due to mishaps on the wiring harness, sensors and circuits were all working as expected.

Guess this sums up the major issues for this sensor hub!

To-Do:
1. Verify duty ratio checking logic, where the sensor hub detect abnormality in ratio of tooth to tooth root.